+++ packages/atproto/lib/src/atproto.dart
@@ -20,6 +20,7 @@ sealed class ATProto {
     final core.Session session, {
     core.Protocol protocol = core.defaultProtocol,
     String service = core.defaultService,
+    String streamService = core.defaultStreamService,
     Duration timeout = core.defaultTimeout,
     core.RetryConfig? retryConfig,
     final core.GetClient? mockedGetClient,
@@ -29,6 +30,7 @@ sealed class ATProto {
         session: session,
         protocol: protocol,
         service: service,
+        streamService: streamService,
         timeout: timeout,
         retryConfig: retryConfig,
         mockedGetClient: mockedGetClient,
@@ -39,6 +41,7 @@ sealed class ATProto {
   factory ATProto.anonymous({
     core.Protocol protocol = core.defaultProtocol,
     String service = core.defaultService,
+    String streamService = core.defaultStreamService,
     Duration timeout = core.defaultTimeout,
     core.RetryConfig? retryConfig,
     final core.GetClient? mockedGetClient,
@@ -47,6 +50,7 @@ sealed class ATProto {
       _ATProto(
         protocol: protocol,
         service: service,
+        streamService: streamService,
         timeout: timeout,
         retryConfig: retryConfig,
         mockedGetClient: mockedGetClient,
@@ -84,6 +88,7 @@ final class _ATProto implements ATProto {
     this.session,
     required core.Protocol protocol,
     required String service,
+    required streamService,
     required Duration timeout,
     core.RetryConfig? retryConfig,
     final core.GetClient? mockedGetClient,
@@ -92,6 +97,7 @@ final class _ATProto implements ATProto {
           did: session?.did ?? '',
           protocol: protocol,
           service: service,
+          streamService: streamService,
           context: core.ClientContext(
             accessJwt: session?.accessJwt ?? '',
             timeout: timeout,