+++ packages/atproto_core/lib/src/const.dart
@@ -16,3 +16,6 @@ const defaultTimeout = Duration(seconds: 30);
 
 /// `bsky.social`
 const defaultService = 'bsky.social';
+
+/// `bsky.network`
+const defaultStreamService = 'bsky.network';