+++ packages/xrpc/lib/src/const.dart
@@ -4,3 +4,6 @@
 
 /// The default service to communicate.
 const defaultService = 'bsky.social';
+
+/// The default service to streamed communicate.
+const defaultStreamService = 'bsky.network';