+++ packages/atproto_core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Note
 
+## v0.6.2
+
+- Fixed a bug that service names were not specified correctly when using stream endpoints.
+
 ## v0.6.1
 
 - Added `BaseHttpService.post`. ([#999](https://github.com/myConsciousness/atproto.dart/issues/999))